feat: impl Eval and Binder traits
*Binder*: this trait is for generically
handling the encoding and decoding of
Var and Lambda.parameter.

*Eval*: this trait is for generically
looking up Var in Env

Eval is implemented on DeBruijn but not
for Name because Name is not evaluatable
there is no way to look it up in the environment.

All Name, NamedDeBruijn, and DeBruijn implement
Binder because they are all technically encodable.

**Note**: only DeBruijn actually ever goes on chain.